# app/docs/swagger.yaml
definitions:
...
  response.ListData:
    properties:
      rows: {}
      total:
        type: integer
    type: object
  response.ResCode:
    enum:
    - 200
    - 401
    - 403
    - 412
    - 500
    - 600
    type: integer
    x-enum-comments:
      Error: 系统异常
      Forbidden: 没有权限
      Parameter: 参数错误
      Success: 成功
      Unauthorized: token失效
      Waring: 详情看msg
    x-enum-varnames:
    - Success
    - Unauthorized
    - Forbidden
    - Parameter
    - Error
    - Waring
  response.ResponseData:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 相应状态码
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
...
  systemModels.SysPermissionAdd:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      parentId:
        description: 父权限ID
        example: "0"
        type: string
      permission:
        description: 权限标识符
        type: string
      permissionName:
        description: 权限名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
    required:
    - parentId
    - permission
    type: object
  systemModels.SysPermissionEdit:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      parentId:
        description: 父权限ID
        example: "0"
        type: string
      permission:
        description: 权限标识符
        type: string
      permissionId:
        description: 权限ID
        example: "0"
        type: string
      permissionName:
        description: 权限名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
    required:
    - permissionId
    type: object
  systemModels.SysPermissionVo:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      parentId:
        description: 父权限ID
        example: "0"
        type: string
      permission:
        description: 权限标识符
        type: string
      permissionId:
        description: 权限ID
        example: "0"
        type: string
      permissionName:
        description: 权限名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
    type: object
  systemModels.SysPostVo:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      postCode:
        type: string
      postId:
        example: "0"
        type: string
      postName:
        type: string
      postSort:
        type: integer
      remark:
        type: string
      status:
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
    type: object
...
info:
  contact:
    email: zhao_402295440@126.com
    name: danny
  description: baize接口文档
  title: baize
  version: 2.0.x
paths:
...
  /system/permission:
    post:
      description: 新增系统权限
      parameters:
      - description: 系统权限
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysPermissionAdd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 新增系统权限
      tags:
      - 系统权限
    put:
      description: 修改系统权限
      parameters:
      - description: 系统权限
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysPermissionEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 修改系统权限
      tags:
      - 系统权限
  /system/permission/{permissionId}:
    delete:
      description: 删除系统权限
      parameters:
      - description: permissionId
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 删除系统权限
      tags:
      - 系统权限
    get:
      description: 根据id查询系统权限
      parameters:
      - description: permissionId
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/systemModels.SysPermissionVo'
              type: object
      security:
      - BearerAuth: []
      summary: 根据id查询系统权限
      tags:
      - 系统权限
  /system/permission/byRoleIds/{roleIds}:
    get:
      description: 根据id查询系统权限
      parameters:
      - description: roleIds
        in: path
        name: roleIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/systemModels.SysPermissionVo'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 根据id查询系统权限
      tags:
      - 系统权限
  /system/permission/list:
    get:
      description: 系统权限列表
      parameters:
      - description: 创建人
        in: query
        name: createBy
        type: integer
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 修改人
        in: query
        name: updateBy
        type: integer
      - description: 修改时间
        in: query
        name: updateTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/systemModels.SysPermissionVo'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 系统权限列表
      tags:
      - 系统权限
  /system/post:
    post:
      description: 添加岗位
      parameters:
      - description: 公司信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysPostVo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 添加岗位
      tags:
      - 岗位相关
    put:
      description: 修改岗位
      parameters:
      - description: 公司信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysPostVo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 修改岗位
      tags:
      - 岗位相关
  /system/post/{postId}:
    get:
      description: 根据岗位ID获取岗位信息
      parameters:
      - description: PostId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/systemModels.SysPostVo'
              type: object
      security:
      - BearerAuth: []
      summary: 根据岗位ID获取岗位信息
      tags:
      - 岗位相关
  /system/post/{postIds}:
    delete:
      description: 删除岗位
      parameters:
      - collectionFormat: csv
        description: postId
        in: path
        items:
          type: string
        name: ids
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 删除岗位
      tags:
      - 岗位相关
  /system/post/export:
    post:
      description: 导出岗位
      parameters:
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: postCode
        type: string
      - in: query
        name: postName
        type: string
      - in: query
        name: status
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
      security:
      - BearerAuth: []
      summary: 导出岗位
      tags:
      - 岗位相关
  /system/post/list:
    get:
      description: 查询岗位列表查询
      parameters:
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: postCode
        type: string
      - in: query
        name: postName
        type: string
      - in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysPostVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询岗位列表查询
      tags:
      - 岗位相关
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"